Fixed the table buttons in the view list panel
Switched the table buttons to rounded buttons and used the cell action listener to invoke actions on the buttons so that it corresponds to the data in that specific row. Added a method to update the students list whenever a search is done. Added a method to use the object in the row of the currently pressed button to set the fields in the update panel to enable editing of data from the view list panel.